Changed implementation
- Implementation now uses 2 grids / arrays - one still of 'Cell' objects,
and another of bytes
- The Cell array is now only responsible for changing state when tasked
to, but the actual live / dead state of the Cells stored here is not
really crucial anymore - will remove that in the next pass
- The byte array is the important part of the implementation.
- Each cell is represented as a byte, with the top 3 bits unused, the
next 4 bits tracking the number of live neighbours the cell has, and the
lowest bit tracking the state (1 = alive; 0 = dead).
- Tracking the number of neighbours this way allows for much faster
processing since it doesn't need to be recalculated if nothing in the
neighbourhood changes. When a cell is toggled, the neighbours are
updated.
- Dead cells with no neighbours can be rapidly skipped during the main
loop.
- Result: almost 3x speed increase; bottle neck is now the rendering